extracted handlePasteEvent into cardsContextMenu and refactored EventContextMenu accordingly
